don't fail if a configured directory doesn't exist

+++ src/DirectoryCleanupCommand.php
@@ -3,6 +3,7 @@
 namespace Spatie\DirectoryCleanup;
 
 use Illuminate\Console\Command;
+use File;
 
 class DirectoryCleanupCommand extends Command
 {
@@ -17,7 +18,9 @@ class DirectoryCleanupCommand extends Command
         $directories = collect(config('laravel-directory-cleanup.directories'));
 
         collect($directories)->each(function ($config, $directory) {
-            $this->deleteFilesIfOlderThanMinutes($directory, $config['deleteAllOlderThanMinutes']);
+            if (File::isDirectory($directory)) {
+                $this->deleteFilesIfOlderThanMinutes($directory, $config['deleteAllOlderThanMinutes']);
+            }
         });
 
         $this->comment('All done!');
